refactor: OrderItem을 domain 레이어로 이동
