+++ app/src/main/java/woowacourse/shopping/domain/model/OrderItem.kt
@@ -1,0 +1,10 @@
+package woowacourse.shopping.domain.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class OrderItem(
+    val cartItemId: Long,
+    val productId: Long,
+    val quantity: Int,
+)
