+++ app/src/main/java/woowacourse/shopping/domain/usecase/BuildPaymentItemsUseCase.kt
@@ -4,9 +4,10 @@ import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.coroutineScope
 import woowacourse.shopping.domain.model.CartItem
+import woowacourse.shopping.domain.model.OrderItem
 import woowacourse.shopping.domain.model.PaymentItems
 import woowacourse.shopping.domain.repository.ProductRepository
-import woowacourse.shopping.presentation.navigation.OrderItem
+import kotlin.collections.map
 
 class BuildPaymentItemsUseCase(
     private val productRepository: ProductRepository,
