+++ app/src/main/java/woowacourse/shopping/presentation/navigation/Route.kt
@@ -1,6 +1,7 @@
 package woowacourse.shopping.presentation.navigation
 
 import kotlinx.serialization.Serializable
+import woowacourse.shopping.domain.model.OrderItem
 
 @Serializable
 object ShoppingScreen
